pyproject.toml
[tool.poetry]
name = "drf-hw"
version = "0.1.0"
description = ""
authors = ["Anna Soboleva <asoboleva.avs@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.5"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
pillow = "^11.1.0"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.4.0"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
coverage = "^7.6.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
